Camada Controls: Adicionado a classe RegisterUserMethods, para que a mesma receba os métodos de seu respectivo form.
